// src/data/pricing-catalogue.generated.ts
export const RATE_CARD: RateCard = {
  labourCostPerPersonHour: 12.6,
  overheadRate: 0.106,
  targetGrossMargin: 0.5,
  defaultMaterialsPerJob: 30,
  defaultTransportPerVisit: 50,
  vatRate: 0.05,
  roundToNearest: 10,
};
// ...
export const ADD_ON_PRICES: PriceRow[] = [
  {
    id: "ADD-01",
    label: "Refrigerator Interior Cleaning",
    attribute: null,
    unit: "Per appliance",
    unitsCovered: 1,
    priceExVat: 99,
    pricePerUnit: 99,
    crew: 1,
    hours: 1,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-02",
    label: "Oven Interior Cleaning",
    attribute: null,
    unit: "Per appliance",
    unitsCovered: 1,
    priceExVat: 69,
    pricePerUnit: 69,
    crew: 1,
    hours: 0.1,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-03",
    label: "Mattress Shampooing",
    attribute: null,
    unit: "Single",
    unitsCovered: 1,
    priceExVat: 129,
    pricePerUnit: 129,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-03",
    label: "Mattress Shampooing",
    attribute: null,
    unit: "Double",
    unitsCovered: 1,
    priceExVat: 159,
    pricePerUnit: 159,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-03",
    label: "Mattress Shampooing",
    attribute: null,
    unit: "Queen",
    unitsCovered: 1,
    priceExVat: 179,
    pricePerUnit: 179,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-03",
    label: "Mattress Shampooing",
    attribute: null,
    unit: "King",
    unitsCovered: 1,
    priceExVat: 229,
    pricePerUnit: 229,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-04",
    label: "Sofa Shampooing",
    attribute: null,
    unit: "Per sofa",
    unitsCovered: 1,
    priceExVat: null,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: "Same price 10 dhs lower",
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-05",
    label: "Carpet Shampooing",
    attribute: null,
    unit: "Per sq.m",
    unitsCovered: 25,
    priceExVat: null,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: "Same price 10 dhs lower",
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-06",
    label: "Intense Deep Cleaning Upgrade",
    attribute: null,
    unit: "Per property",
    unitsCovered: 1,
    priceExVat: null,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: "Wtv is the difference in the standard services",
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-07",
    label: "Interior Cabinet Cleaning",
    attribute: null,
    unit: "Per property",
    unitsCovered: 1,
    priceExVat: 99,
    pricePerUnit: 99,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "ADD-11",
    label: "Curtain Cleaning & Shampooing",
    attribute: null,
    unit: "Per panel",
    unitsCovered: 10,
    priceExVat: null,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: "On qoute",
    quoteOnly: true,
    aliasOf: null,
  },
];

export const MANPOWER_PRICES: PriceRow[] = [
  {
    id: "SVC-501",
    label: "Office Boy",
    attribute: null,
    unit: "Per month",
    unitsCovered: null,
    priceExVat: 3799,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "SVC-502",
    label: "Cleaner",
    attribute: null,
    unit: "Per month",
    unitsCovered: null,
    priceExVat: 3499,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
  {
    id: "SVC-504",
    label: "Watchman",
    attribute: null,
    unit: "Per month",
    unitsCovered: null,
    priceExVat: 4199,
    pricePerUnit: null,
    crew: null,
    hours: null,
    note: null,
    quoteOnly: false,
    aliasOf: null,
  },
];